+++ audio-streaming-client-cpp/demo/client_demo_c.cc
@@ -14,7 +14,7 @@ void default_callback(AudioFragmentResponseWrapper* resp, void* data) {
 typedef struct {
     AsrStreamWrapper* stream;
     FILE* fp;
-    int send_package_size;
+    unsigned int send_package_size;
 } WriteArguments;
 
 void* write_to_stream(void* args) {
@@ -56,7 +56,7 @@ int main(int argc, char* argv[]) {
     // create AsrClient
     AsrClientWrapper* client = asr_client_create();
 
-    asr_client_set_flush_data(client, true);
+    asr_client_set_enable_flush_data(client, true);
     if (argc == 3) {
         asr_client_set_product_id(client, argv[1]);
         if (asr_client_init(client, argv[2]) != 0) {
@@ -86,7 +86,7 @@ int main(int argc, char* argv[]) {
     WriteArguments arguments;
     arguments.stream = stream_wrapper;
     arguments.fp = fp;
-    arguments.send_package_size = 320;
+    arguments.send_package_size = asr_client_get_send_package_size(client);
 
     pthread_create(
             &writer,