+++ audio-streaming-client-java/src/main/java/com/baidu/acu/pie/demo/JavaDemo.java
@@ -2,24 +2,24 @@
 
 package com.baidu.acu.pie.demo;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.time.Instant;
-import java.util.List;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-
 import com.baidu.acu.pie.AudioStreaming.AudioFragmentRequest;
 import com.baidu.acu.pie.client.AsrClient;
 import com.baidu.acu.pie.client.AsrClientFactory;
+import com.baidu.acu.pie.client.Consumer;
 import com.baidu.acu.pie.model.AsrConfig;
 import com.baidu.acu.pie.model.AsrProduct;
 import com.baidu.acu.pie.model.RecognitionResult;
 import com.google.protobuf.ByteString;
-
 import io.grpc.stub.StreamObserver;
+import org.joda.time.DateTime;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.List;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 
 /**
  * JavaDemo
@@ -45,7 +45,7 @@ public class JavaDemo {
     }
 
     public void recognizeFile() {
-        String audioFilePath = "testaudio/1.wav";
+        String audioFilePath = "testaudio/bj8k.wav";
         AsrClient asrClient = createAsrClient();
         List<RecognitionResult> results = asrClient.syncRecognize(Paths.get(audioFilePath));
 
@@ -54,6 +54,7 @@ public class JavaDemo {
 
         for (RecognitionResult result : results) {
             System.out.println(String.format(AsrConfig.TITLE_FORMAT,
+                    "file_name",
                     "serial_num",
                     "err_no",
                     "err_message",
@@ -61,6 +62,7 @@ public class JavaDemo {
                     "end_time",
                     "result"));
             System.out.println(String.format(AsrConfig.TITLE_FORMAT,
+                    audioFilePath,
                     result.getSerialNum(),
                     result.getErrorCode(),
                     result.getErrorMessage(),
@@ -75,29 +77,37 @@ public class JavaDemo {
      * 使用长音频来模拟人对着麦克风不断说话的情况
      */
     public void asyncRecognition() {
-        String longAudioFilePath = "testaudio/1.wav";
+        String longAudioFilePath = "testaudio/300s.wav";
         AsrClient asrClient = createAsrClient();
         CountDownLatch finishLatch = new CountDownLatch(1);
-        StreamObserver<AudioFragmentRequest> sender = asrClient.asyncRecognize(it -> {
-            System.out.println(
-                    Instant.now().toString() + "\t" + Thread.currentThread().getId() + " receive fragment: " + it);
+        StreamObserver<AudioFragmentRequest> sender = asrClient.asyncRecognize(new Consumer<RecognitionResult>() {
+            @Override
+            public void accept(RecognitionResult it) {
+                System.out.println(
+//                        Instant.now().toString() + "\t" + Thread.currentThread().getId() + " receive fragment: " + it);
+                        Thread.currentThread().getId() + " receive fragment: " + it);
+            }
         }, finishLatch);
 
+        // 这里从文件中得到一个InputStream，实际场景下，也可以从麦克风或者其他音频源来得到InputStream
         try (InputStream audioStream = Files.newInputStream(Paths.get(longAudioFilePath))) {
             byte[] data = new byte[asrClient.getFragmentSize()];
             int readSize;
 
-            System.out.println(Instant.now().toString() + "\t" + Thread.currentThread().getId() + " start to send");
+            System.out.println(new DateTime().toString() + "\t" + Thread.currentThread().getId() + " start to send");
+
+            // 使用 sender.onNext 方法，将 InputStream 中的数据不断地发送到 asr 后端，发送的最小单位是 AudioFragment
             while ((readSize = audioStream.read(data)) != -1) {
-                System.out.println(Instant.now().toString() + "\t"
-                        + Thread.currentThread().getId() + " send fragment");
+                System.out.println(new DateTime().toString() + "\t"
+                        +Thread.currentThread().getId() + " send fragment");
                 sender.onNext(AudioFragmentRequest.newBuilder()
                         .setAudioData(ByteString.copyFrom(data, 0, readSize))
                         .build());
                 // 主动休眠一段时间，来模拟人说话场景下的音频产生速率
+                // 在对接麦克风等设备的时候，可以去掉这个 sleep
                 Thread.sleep(20);
             }
-            System.out.println(Instant.now().toString() + "\t" + Thread.currentThread().getId() + " send finish");
+            System.out.println(new DateTime().toString() + "\t" + Thread.currentThread().getId() + " send finish");
 
             // wait to ensure to receive the last response
             finishLatch.await(1000, TimeUnit.SECONDS);